Update device integration library to support new PowerMon firmware version
Adapt the PowerMon library and bridge to accommodate changes in version 1.12, including removing the deprecated `initBle()` call, updating the `DeviceInfo` structure to include MAC address, and adjusting schedule repeat type.

+++ libpowermon_bin_v117_backup/inc/powermon_schedule.h
@@ -57,8 +57,8 @@ public:
 	bool isRepeatDOM(void) const;
 	void setRepeatDOM(void);
 
-	uint32_t getRepeat(void) const;
-	void setRepeat(uint32_t value);
+	uint8_t getRepeat(void) const;
+	void setRepeat(uint8_t value);
 
 private:
 	uint8_t mRawSchedule[24];
